Operator = (char*) added, atribut length_ (redundant with size_) deleted
Operator = (char*)
Assigns a new value to the string, replacing its current contents with
a pointer to a null-terminated sequence of characters passed as parameter.
The sequence is copied as the new value for the string.
length_
This atribut was redundant with size_ (they had the same value and
represented the same thing). Howether we kept the length() methode, which
return size_

Signed-off-by: Titouan <tpoquillon@gmail.com>